+++ CHANGELOG.md
@@ -3,6 +3,7 @@ In this file I'll try to mention major changes to the configuration.
 
 2025-12-13
 - hardware: fixed the email address for smartd reports.
+- repo: moved github runner from contabo1 to hostup1
 
 2025-12-12
 - laptop profile: added font package containing a cjk font.